refactor: streamline MessageManager component and remove unused files
- Refactored the MessageManager component to improve structure and readability.
- Integrated new SuggestionBox and TextBox components for enhanced functionality.
- Removed obsolete components such as FileUploader, LoadingBar, and MoreActionsOption to reduce clutter.
- Updated styles and logic for better user experience and performance.
