+++ src/components/chats/MessageManager/SuggestionBox/SuggestionBoxShortcut.vue
