+++ src/components/chats/MessageManager/TextBox/TextArea.vue
