+++ src/components/chats/MessageManager/index.vue
@@ -1,582 +1,153 @@
 <template>
-  <section>
-    <MessageManagerLoading v-show="showSkeletonLoading" />
-    <div
-      v-show="!showSkeletonLoading"
-      class="message-manager"
-      data-testid="message-manager"
-    >
-      <div
-        :class="[
-          'message-manager-box__container',
-          isFileLoadingValueValid && 'loading',
-          isFocused && 'focused',
-        ]"
-        data-testid="message-manager-box-container"
-      >
-        <LoadingBar
-          v-if="isFileLoadingValueValid"
-          :value="loadingFileValue"
-        />
-        <UnnnicReplyMessage
-          v-if="replyMessage"
-          class="message-manager__reply-message"
-          :replyMessage="replyMessage"
-          showClose
-          messageType="received"
-          @close="clearReplyMessage()"
-        />
-
-        <TextBox
-          v-if="!isAudioRecorderVisible"
-          ref="textBox"
-          class="message-manager-box__text-box"
-          :modelValue="inputMessage"
-          :isInternalNote="isInternalNote"
-          @update:model-value="onInputMessageUpdate"
-          @keydown.stop="onKeyDown"
-          @paste="handlePaste"
-          @is-typing-handler="isTypingHandler"
-          @is-focused-handler="isFocusedHandler"
-          @handle-quick-messages="emitShowQuickMessages"
-          @open-file-uploader="openFileUploader"
-          @close-internal-note="handleInternalNoteInput"
-        />
-        <UnnnicAudioRecorder
-          v-show="isAudioRecorderVisible && !isFileLoadingValueValid"
-          ref="audioRecorder"
-          v-model="audioMessage"
-          class="message-manager__audio-recorder"
-          @status="updateAudioRecorderStatus"
-        />
-      </div>
-      <div class="message-manager__actions">
-        <UnnnicButton
-          v-if="
-            canUseCopilot && !isCopilotOpen && showActionButton && !discussionId
-          "
-          type="secondary"
-          size="large"
-          iconCenter="wb_incandescent"
-          class="message-manager__actions__co-pilot"
-          @click="openCopilot"
-        />
-        <UnnnicButton
-          v-if="
-            (!canUseCopilot || discussionId) &&
-            showActionButton &&
-            !isInternalNote
-          "
-          type="secondary"
-          size="large"
-          iconCenter="mic"
-          @click="record"
-        />
-
-        <UnnnicButton
-          v-if="discussionId && showActionButton"
-          type="secondary"
-          size="large"
-          iconCenter="attachment"
-          next
-          @click="openFileUploader"
-        />
-
-        <UnnnicDropdown
-          v-if="
-            (showActionButton || isSuggestionBoxOpen) &&
-            !discussionId &&
-            !isInternalNote
-          "
-          position="top-left"
-          class="more-actions"
-        >
-          <template #trigger>
-            <UnnnicButton
-              type="primary"
-              size="large"
-              iconCenter="add"
-            />
-          </template>
-
-          <div class="more-actions-container">
-            <MoreActionsOption
-              v-if="canUseCopilot"
-              :action="record"
-              icon="mic"
-              :title="$t('record_audio')"
-            />
-            <MoreActionsOption
-              :action="openFileUploader"
-              icon="attachment"
-              :title="$t('attach')"
-            />
-            <MoreActionsOption
-              :action="emitShowQuickMessages"
-              icon="bolt"
-              :title="$t('quick_message')"
-            />
-            <MoreActionsOption
-              :action="handleInternalNoteInput"
-              icon="add_notes"
-              :title="$t('internal_note')"
-            />
-          </div>
-        </UnnnicDropdown>
-
-        <UnnnicButton
-          v-if="showSendMessageButton"
-          :type="isInternalNote ? 'attention' : 'primary'"
-          size="large"
-          iconCenter="send"
-          class="send-message-button"
-          @click="send"
-        />
-        <UnnnicButton
-          v-else-if="isMobile"
-          type="primary"
-          size="large"
-          iconCenter="mic"
-          @click="record"
-        />
-      </div>
-      <SuggestionBox
-        v-if="!discussionId"
-        :search="inputMessage"
-        :suggestions="shortcuts"
-        :keyboardEvent="keyboardEvent"
-        :copilot="canUseCopilot && !discussionId"
-        @open="isSuggestionBoxOpen = true"
-        @close="closeSuggestionBox"
-        @select="setMessage($event.text)"
-        @open-copilot="openCopilot"
-      />
-      <CoPilot
-        v-if="isCopilotOpen"
-        ref="copilot"
-        @select="setMessage($event)"
-        @close="isCopilotOpen = false"
-      />
-    </div>
+  <section class="message-manager-v2">
+    <MessageManagerLoading v-if="isLoading" />
+    <MessageManagerTextBox
+      v-else
+      ref="messageManagerTextBox"
+      @keydown="onKeyDown"
+    />
+    <SuggestionBox
+      v-if="!activeDiscussion?.uuid"
+      :search="inputMessage"
+      :keyboardEvent="keyboardEvent"
+      :ignoreClickOutside="suggestionBoxIgnoreClickOutside"
+      @open="isSuggestionBoxOpen = true"
+      @hide="isSuggestionBoxOpen = false"
+      @close="closeSuggestionBox"
+      @select="(text) => setMessage(text)"
+    />
+    <CoPilot
+      v-if="isCopilotOpen"
+      @select="(text) => setMessage(text)"
+    />
   </section>
 </template>
 
-<script>
-import isMobile from 'is-mobile';
+<script setup lang="ts">
+import { nextTick, onMounted, ref, useTemplateRef, watch } from 'vue';
+import { storeToRefs } from 'pinia';
 
-import { mapActions, mapState, mapWritableState } from 'pinia';
-
-import { useQuickMessageShared } from '@/store/modules/chats/quickMessagesShared';
-import { useQuickMessages } from '@/store/modules/chats/quickMessages';
-import { useRooms } from '@/store/modules/chats/rooms';
-import { useDiscussions } from '@/store/modules/chats/discussions';
-import { useDiscussionMessages } from '@/store/modules/chats/discussionMessages';
-import { useRoomMessages } from '@/store/modules/chats/roomMessages';
-import { useFeatureFlag } from '@/store/modules/featureFlag';
-import { useMessageManager } from '@/store/modules/chats/messageManager';
-
-import MessageManagerLoading from '@/views/loadings/chat/MessageManager.vue';
-
-import TextBox from './TextBox.vue';
-import MoreActionsOption from './MoreActionsOption.vue';
-import LoadingBar from './LoadingBar.vue';
-import SuggestionBox from './SuggestionBox.vue';
+import MessageManagerLoading from './MessageManagerLoading.vue';
+import MessageManagerTextBox from './TextBox/index.vue';
+import SuggestionBox from './SuggestionBox/index.vue';
 import CoPilot from './CoPilot.vue';
 
-export default {
-  name: 'MessageManager',
+import { useDiscussions } from '@/store/modules/chats/discussions';
+import { useMessageManager } from '@/store/modules/chats/messageManager';
+import { useRooms } from '@/store/modules/chats/rooms';
+import { useProfile } from '@/store/modules/profile';
 
-  components: {
-    MessageManagerLoading,
-    TextBox,
-    LoadingBar,
-    SuggestionBox,
-    MoreActionsOption,
-    CoPilot,
-  },
+defineOptions({
+  name: 'MessageManagerV2',
+});
 
-  props: {
-    loadingFileValue: {
-      validator: (value) => {
-        return value === null || typeof value === 'number';
-      },
-      default: null,
-    },
-    showSkeletonLoading: {
-      type: Boolean,
-      default: false,
-    },
-  },
-  emits: ['show-quick-messages', 'open-file-uploader'],
+interface Props {
+  isLoading: boolean;
+}
 
-  data: () => ({
-    keyboardEvent: null,
-    isSuggestionBoxOpen: false,
-    isCopilotOpen: false,
-    isTyping: false,
-    isFocused: false,
-    isLoading: false,
-    isInternalNote: false,
-  }),
+const props = defineProps<Props>();
 
-  computed: {
-    ...mapState(useQuickMessageShared, ['quickMessagesShared']),
-    ...mapState(useQuickMessages, ['quickMessages']),
-    ...mapState(useRooms, ['canUseCopilot', 'activeRoom']),
-    ...mapState(useDiscussions, {
-      discussionId: (store) => store.activeDiscussion?.uuid,
-    }),
-    ...mapWritableState(useMessageManager, [
-      'inputMessage',
-      'audioMessage',
-      'audioRecorderStatus',
-      'replyMessage',
-    ]),
-    ...mapState(useFeatureFlag, ['featureFlags']),
+const profileStore = useProfile();
+const { me } = storeToRefs(profileStore);
 
-    isMobile() {
-      return isMobile();
-    },
+const roomsStore = useRooms();
+const { activeRoom } = storeToRefs(roomsStore);
 
-    isAudioRecorderVisible() {
-      return (
-        !!this.audioMessage ||
-        ['recording', 'recorded', 'playing', 'paused'].includes(
-          this.audioRecorderStatus,
-        )
-      );
-    },
-    isFileLoadingValueValid() {
-      return typeof this.loadingFileValue === 'number';
-    },
-    shortcuts() {
-      const allShortcuts = [...this.quickMessages, ...this.quickMessagesShared];
-      const uniqueShortcuts = [];
+const discussionsStore = useDiscussions();
+const { activeDiscussion } = storeToRefs(discussionsStore);
 
-      allShortcuts.forEach((item) => {
-        const isDuplicate = uniqueShortcuts.some(
-          (uniqueItem) => uniqueItem.uuid === item.uuid,
-        );
+const messageManager = useMessageManager();
+const { sendRoomMessage, sendMediasMessage } = messageManager;
+const {
+  inputMessage,
+  isCopilotOpen,
+  isSuggestionBoxOpen,
+  isInternalNote,
+  mediaUploadFiles,
+  inputMessageFocused,
+} = storeToRefs(messageManager);
 
-        if (!isDuplicate) {
-          uniqueShortcuts.push(item);
-        }
-      });
+const keyboardEvent = ref<KeyboardEvent | null>(null);
+const messageManagerTextBoxRef = useTemplateRef('messageManagerTextBox');
 
-      return uniqueShortcuts;
-    },
-    showActionButton() {
-      const {
-        isTyping,
-        isAudioRecorderVisible,
-        isFileLoadingValueValid,
-        isMobile,
-      } = this;
-      return (
-        !isTyping &&
-        !isAudioRecorderVisible &&
-        !isFileLoadingValueValid &&
-        !isMobile
-      );
-    },
-    showSendMessageButton() {
-      const {
-        isSuggestionBoxOpen,
-        isTyping,
-        isAudioRecorderVisible,
-        isFileLoadingValueValid,
-        isInternalNote,
-      } = this;
-      return (
-        !isSuggestionBoxOpen &&
-        (isTyping ||
-          isAudioRecorderVisible ||
-          isFileLoadingValueValid ||
-          isInternalNote)
-      );
-    },
-  },
+const suggestionBoxIgnoreClickOutside = [messageManagerTextBoxRef];
 
-  watch: {
-    'activeRoom.uuid'() {
-      this.clearReplyMessage();
-    },
-    replyMessage(newReplyMessage) {
-      if (newReplyMessage) this.focusTextBox();
-    },
-  },
-
-  mounted() {
-    this.clearReplyMessage();
-  },
-
-  methods: {
-    ...mapActions(useDiscussionMessages, [
-      'sendDiscussionMessage',
-      'sendDiscussionMedias',
-    ]),
-    ...mapActions(useRoomMessages, [
-      'sendRoomMessage',
-      'sendRoomMedias',
-      'sendRoomInternalNote',
-    ]),
-    handleInternalNoteInput() {
-      this.clearTextBox();
-      this.clearReplyMessage();
-      this.isInternalNote = !this.isInternalNote;
-    },
-    onInputMessageUpdate(val) {
-      this.inputMessage = val;
-    },
-    openCopilot() {
-      this.isCopilotOpen = true;
-      this.clearTextBox();
-    },
-    emitShowQuickMessages() {
-      this.$emit('show-quick-messages');
-    },
-    setMessage(newMessage) {
-      this.inputMessage = newMessage;
-      this.$nextTick(() => {
-        this.focusTextBox();
-      });
-    },
-    clearReplyMessage() {
-      this.replyMessage = null;
-    },
-    clearAudio() {
-      this.$refs.audioRecorder?.discard();
-      this.audioMessage = null;
-    },
-    clearTextBox() {
-      this.inputMessage = '';
-    },
-    /**
-     * @param {KeyboardEvent} event
-     */
-    closeSuggestionBox() {
-      this.isSuggestionBoxOpen = false;
-
-      if (this.inputMessage.startsWith('/')) {
-        this.inputMessage = '';
-      }
-    },
-    onKeyDown(event) {
-      if (this.isSuggestionBoxOpen) {
-        if (event.key === 'Escape') {
-          this.closeSuggestionBox();
-          return;
-        }
-        this.keyboardEvent = event;
-        return;
-      }
-
-      if (event.key === 'Enter') {
-        if (event.shiftKey) return;
-        this.send();
-        event.preventDefault();
-      }
-
-      if (event.key === 'Escape' && this.isInternalNote) {
-        this.handleInternalNoteInput();
-        event.preventDefault();
-      }
-    },
-    isTypingHandler(isTyping = false) {
-      this.isTyping = isTyping;
-    },
-    isFocusedHandler(isFocused = false) {
-      this.isFocused = isFocused;
-    },
-    handlePaste(event) {
-      const { items } =
-        event.clipboardData || event.originalEvent.clipboardData;
-      const itemsArray = [...items];
-      const imagePastes = itemsArray.filter(
-        (item) => item.type.includes('image') || item.type === 'video/mp4',
-      );
-
-      const fileList = imagePastes.map((imagePaste) => {
-        const blob = imagePaste.getAsFile();
-        const dateOfPrintPaste = new Date(Date.now()).toUTCString();
-        const fileName =
-          blob.name === 'image.png' ? `${dateOfPrintPaste}.png` : blob.name;
-        const file = new File([blob], fileName, { type: blob.type });
-        return file;
-      });
-
-      if (fileList.length) {
-        this.openFileUploader(fileList);
-      }
-    },
-    record() {
-      if (!this.isLoading) {
-        this.$refs.audioRecorder?.record();
-      } else {
-        console.info('Loading');
-      }
-    },
-    stopRecord() {
-      this.$refs.audioRecorder?.stop();
-    },
-    async send() {
-      if (this.isInternalNote) {
-        await this.sendInternalNote();
-      } else {
-        let repliedMessage = null;
-        if (this.replyMessage) {
-          repliedMessage = { ...this.replyMessage };
-          this.replyMessage = null;
-        }
-        await this.sendTextBoxMessage(repliedMessage);
-        await this.sendAudio(repliedMessage);
-      }
-      this.$refs.textBox?.clearTextarea();
-    },
-    async sendInternalNote() {
-      if (!this.inputMessage.trim()) return;
-      const text = `${this.$t('internal_note')}: ${this.inputMessage.trim()}`;
-      this.clearTextBox();
-      await this.sendRoomInternalNote({ text });
-      this.handleInternalNoteInput();
-    },
-    async sendTextBoxMessage(repliedMessage) {
-      const message = this.inputMessage.trim();
-      if (message) {
-        this.clearTextBox();
-        if (this.discussionId) {
-          await this.sendDiscussionMessage(message);
-        } else {
-          await this.sendRoomMessage(message, repliedMessage);
-        }
-      }
-    },
-    async sendAudio(repliedMessage) {
-      if (this.audioRecorderStatus === 'recording') {
-        await this.stopRecord();
-      }
-
-      if (!this.audioMessage || this.isLoading) return;
-      this.isLoading = true;
-
-      const loadingFiles = {};
-      const updateLoadingFiles = (messageUuid, progress) => {
-        loadingFiles[messageUuid] = progress;
-        this.totalValue =
-          Object.values(loadingFiles).reduce((acc, value) => acc + value) /
-          Object.keys(loadingFiles).length;
-      };
-      const response = await fetch(this.audioMessage.src);
-      const blob = await response.blob();
-      const audio = new File([blob], `${Date.now().toString()}.mp3`, {
-        type: 'audio/mpeg3',
-      });
-
-      const sendPayload = {
-        files: [audio],
-        updateLoadingFiles,
-        repliedMessage,
-      };
-
-      if (this.discussionId) {
-        await this.sendDiscussionMedias(sendPayload);
-      } else {
-        await this.sendRoomMedias(sendPayload);
-      }
-
-      this.totalValue = undefined;
-      this.clearAudio();
-
-      this.isLoading = false;
-    },
-    openFileUploader(files, filesType) {
-      this.$emit(
-        'open-file-uploader',
-        files?.length > 0 ? files : [],
-        filesType,
-      );
-    },
-    updateAudioRecorderStatus(status) {
-      this.audioRecorderStatus = status;
-    },
-    focusTextBox() {
-      this.$refs.textBox.focus();
-    },
-  },
+const focusMessageManagerTextBox = () => {
+  nextTick(() => {
+    messageManagerTextBoxRef.value?.focus();
+  });
 };
+
+const setMessage = (message: string) => {
+  inputMessage.value = message || '';
+
+  if (isCopilotOpen.value) {
+    isCopilotOpen.value = false;
+  }
+
+  focusMessageManagerTextBox();
+};
+
+const onKeyDown = (event: KeyboardEvent) => {
+  if (isSuggestionBoxOpen.value) {
+    if (event.key === 'Escape') {
+      isSuggestionBoxOpen.value = false;
+      inputMessage.value = '';
+      return;
+    }
+    keyboardEvent.value = event;
+    return;
+  }
+
+  if (event.key === 'Escape' && isInternalNote.value) {
+    isInternalNote.value = false;
+    event.preventDefault();
+  }
+
+  if (event.key === 'Enter') {
+    if (event.shiftKey) return;
+    if (mediaUploadFiles.value.length > 0) {
+      sendMediasMessage();
+    } else {
+      sendRoomMessage();
+    }
+    event.preventDefault();
+  }
+};
+
+const closeSuggestionBox = () => {
+  isSuggestionBoxOpen.value = false;
+  inputMessage.value = '';
+};
+
+const handleOpenMeActiveRoom = () => {
+  const isMeActiveRoom = activeRoom.value?.user?.email === me.value?.email;
+  if (isMeActiveRoom && !inputMessageFocused.value) {
+    focusMessageManagerTextBox();
+  }
+};
+
+watch(
+  () => props.isLoading,
+  () => {
+    if (!props.isLoading && !inputMessageFocused.value) {
+      focusMessageManagerTextBox();
+    }
+  },
+);
+
+onMounted(() => {
+  handleOpenMeActiveRoom();
+});
 </script>
 
 <style lang="scss" scoped>
-.message-manager {
+.message-manager-v2 {
   position: relative;
 
   display: grid;
   grid-template-columns: 1fr auto;
-  gap: $unnnic-spacing-stack-xs;
+  gap: $unnnic-space-2;
   align-items: end;
-
-  :deep(.reply-message) {
-    box-shadow: 0px 2px 5px -1px rgba(0, 0, 0, 0.1);
-  }
-
-  &-box {
-    &__container {
-      position: relative;
-
-      height: 100%;
-
-      display: flex;
-      flex-direction: column;
-      gap: $unnnic-spacing-nano;
-
-      &.focused {
-        border-color: $unnnic-color-border-soft;
-      }
-
-      &.loading {
-        border-radius: 0 0 $unnnic-border-radius-sm $unnnic-border-radius-sm;
-      }
-
-      &.recording {
-        border: 1px solid $unnnic-color-border-soft;
-        border-radius: $unnnic-border-radius-sm;
-        background-color: $unnnic-color-bg-base;
-      }
-    }
-  }
-
-  &__actions {
-    margin-right: $unnnic-spacing-inline-sm;
-
-    display: flex;
-    gap: $unnnic-spacing-stack-xs;
-
-    &__co-pilot {
-      rotate: 180deg;
-    }
-
-    .more-actions {
-      :deep(.unnnic-dropdown__content) {
-        padding: 0 $unnnic-spacing-inset-sm;
-      }
-    }
-  }
-
-  &__audio-recorder {
-    width: 100%;
-    height: 100%;
-
-    justify-content: flex-end;
-
-    padding: 10px;
-
-    border: 1px solid $unnnic-color-border-soft;
-    border-radius: $unnnic-border-radius-sm;
-    background-color: $unnnic-color-bg-base;
-
-    :deep(.audio-player) {
-      width: auto;
-    }
-  }
 }
 </style>
