+++ src/views/chats/Home/HomeChatModals.vue
@@ -36,17 +36,6 @@
       :room="room"
       @close="closeModal('closeChat')"
     />
-
-    <FileUploader
-      v-if="!enableMessageManagerV2"
-      ref="fileUploader"
-      v-model="modalFileUploaderFiles"
-      :mediasType="modalFileUploaderMediaType"
-      data-testid="modal-file-uploader"
-      @progress="emitFileUploaderProgress"
-      @close="closeModal('fileUploader')"
-      @update:model-value="modalFileUploaderFiles = $event"
-    />
   </section>
 </template>
 
@@ -57,7 +46,6 @@ import isMobile from 'is-mobile';
 import { useRooms } from '@/store/modules/chats/rooms';
 import { useDashboard } from '@/store/modules/dashboard';
 
-import FileUploader from '@/components/chats/MessageManager/FileUploader.vue';
 import ModalGetChat from '@/components/chats/chat/ModalGetChat.vue';
 
 import ModalCloseChat from './ModalCloseChat.vue';
@@ -68,7 +56,6 @@ export default {
   name: 'HomeChatModals',
 
   components: {
-    FileUploader,
     ModalGetChat,
     ModalCloseChat,
   },
@@ -82,7 +69,6 @@ export default {
         getChat: false,
         assumedChat: false,
         closeChat: false,
-        fileUploader: false,
       },
 
       modalFileUploaderMediaType: '',
@@ -99,21 +85,6 @@ export default {
     ...mapWritableState(useMessageManager, {
       modalFileUploaderFiles: 'mediaUploadFiles',
     }),
-    enableMessageManagerV2() {
-      return this.featureFlags.active_features?.includes(
-        'weniChatsInputMessageV2',
-      );
-    },
-  },
-
-  watch: {
-    'modalsShowing.fileUploader': {
-      handler(newModalsShowingFileUploader) {
-        if (newModalsShowingFileUploader && !this.enableMessageManagerV2) {
-          this.$refs.fileUploader.open();
-        }
-      },
-    },
   },
 
   methods: {
@@ -130,13 +101,6 @@ export default {
     },
     closeModal(modalName) {
       this.toggleModal(modalName, 'close');
-    },
-
-    configFileUploader({ files, filesType }) {
-      this.addMediaUploadFiles(files);
-      if (filesType) {
-        this.modalFileUploaderMediaType = filesType;
-      }
     },
 
     emitGotChat() {
